Redesign, composer package
