+++ bootstrap.php
@@ -1,11 +1,5 @@
 <?php
 
-
 Autoloader::add_classes(array(
-	/**
-	 * Email class.
-	 */
 	'Email\\Email_Driver_Mandrill' => __DIR__.'/classes/email/driver/mandrill.php',
-    
-	'Mandrill' => __DIR__.'/vendor/Mandrill.php',
 ));
